+++ partner-chains-cli/src/deregister/tests.rs
@@ -41,9 +41,6 @@ fn errors_if_smart_contracts_dont_output_transaction_id() {
 			read_keys_io(),
 			establish_pc_contracts_cli_configuration_io(None, PcContractsCliResources::default()),
 			run_smart_contract_io(r#"(NotFoundInputUtxo "Couldn't find registration UTxO")"#),
-			MockIO::print(
-				r#"Deregistration failed: (NotFoundInputUtxo "Couldn't find registration UTxO")"#,
-			),
 		]);
 	let result = DeregisterCmd.run(&mock_context);
 	assert_eq!(
@@ -56,16 +53,7 @@ fn errors_if_smart_contracts_dont_output_transaction_id() {
 fn fails_when_chain_config_is_not_valid() {
 	let mock_context = MockIOContext::new()
 		.with_json_file(CHAIN_CONFIG_FILE_PATH, invalid_chain_config_content())
-		.with_expected_io(vec![
-			read_config_twice_io(),
-			print_info_io(),
-			read_keys_io(),
-			establish_pc_contracts_cli_configuration_io(None, PcContractsCliResources::default()),
-			run_smart_contract_io(r#"(NotFoundInputUtxo "Couldn't find registration UTxO")"#),
-			MockIO::print(
-				r#"Deregistration failed: (NotFoundInputUtxo "Couldn't find registration UTxO")"#,
-			),
-		]);
+		.with_expected_io(vec![read_config_twice_io()]);
 	let result = DeregisterCmd.run(&mock_context);
 	assert_eq!(
 	    result.err().unwrap().to_string(),
@@ -190,6 +178,13 @@ fn test_chain_config_content() -> serde_json::Value {
 			"committee_candidates_address": "addr_test1wz5qc7fk2pat0058w4zwvkw35ytptej3nuc3je2kgtan5dq3rt4sc",
 			"d_parameter_policy_id": "d0ebb61e2ba362255a7c4a253c6578884603b56fb0a68642657602d6",
 			"permissioned_candidates_policy_id": "58b4ba68f641d58f7f1bba07182eca9386da1e88a34d47a14638c3fe",
+			"native_token": {
+				"asset": {
+					"policy_id": "ada83ddd029614381f00e28de0922ab0dec6983ea9dd29ae20eef9b4",
+					"asset_name": "5043546f6b656e44656d6f",
+				},
+				"illiquid_supply_address": "addr_test1wrhvtvx3f0g9wv9rx8kfqc60jva3e07nqujk2cspekv4mqs9rjdvz"
+			},
 		},
 		"initial_permissioned_candidates": []
 	})