+++ partner-chains-cli/src/setup_main_chain_state/tests.rs
@@ -29,8 +29,8 @@ fn no_ariadne_parameters_on_main_chain_no_updates() {
 			print_post_update_info_io(),
 		]);
 	let result = SetupMainChainStateCmd.run(&mock_context);
-	mock_context.no_more_io_expected();
-	assert!(result.is_ok());
+
+	result.expect("should succeed");
 }
 
 #[test]
@@ -52,8 +52,7 @@ fn no_ariadne_parameters_on_main_chain_do_updates() {
 			print_post_update_info_io(),
 		]);
 	let result = SetupMainChainStateCmd.run(&mock_context);
-	mock_context.no_more_io_expected();
-	assert!(result.is_ok());
+	result.expect("should succeed");
 }
 
 #[test]
@@ -74,8 +73,7 @@ fn ariadne_parameters_are_on_main_chain_no_updates() {
 			print_post_update_info_io(),
 		]);
 	let result = SetupMainChainStateCmd.run(&mock_context);
-	mock_context.no_more_io_expected();
-	assert!(result.is_ok());
+	result.expect("should succeed");
 }
 
 #[test]
@@ -98,8 +96,7 @@ fn ariadne_parameters_are_on_main_chain_do_update() {
 			print_post_update_info_io(),
 		]);
 	let result = SetupMainChainStateCmd.run(&mock_context);
-	mock_context.no_more_io_expected();
-	assert!(result.is_ok());
+	result.expect("should succeed");
 }
 
 #[test]
@@ -118,8 +115,7 @@ fn fails_if_update_permissioned_candidates_fail() {
 			update_permissioned_candidates_failed_io(),
 		]);
 	let result = SetupMainChainStateCmd.run(&mock_context);
-	assert!(result.is_err());
-	mock_context.no_more_io_expected();
+	result.expect_err("should return error");
 }
 
 #[test]
@@ -139,8 +135,7 @@ fn candidates_on_main_chain_are_same_as_in_config_no_updates() {
 			print_post_update_info_io(),
 		]);
 	let result = SetupMainChainStateCmd.run(&mock_context);
-	mock_context.no_more_io_expected();
-	assert!(result.is_ok());
+	result.expect("should succeed");
 }
 
 #[test]
@@ -149,9 +144,9 @@ fn should_return_error_message_if_pc_cli_missing() {
 		.with_json_file(CHAIN_CONFIG_FILE_PATH, test_chain_config_content())
 		.with_expected_io(vec![read_chain_config_io(), print_info_io()]);
 	let result = SetupMainChainStateCmd.run(&mock_context);
-	mock_context.no_more_io_expected();
+	let err = result.expect_err("should return error");
 	assert_eq!(
-		result.unwrap_err().to_string(),
+		err.to_string(),
 		"Partner Chains Smart Contracts executable file (./pc-contracts-cli) is missing"
 	);
 }
@@ -353,6 +348,13 @@ fn test_chain_config_content() -> serde_json::Value {
 			"committee_candidates_address": "addr_test1wz5qc7fk2pat0058w4zwvkw35ytptej3nuc3je2kgtan5dq3rt4sc",
 			"d_parameter_policy_id": "d0ebb61e2ba362255a7c4a253c6578884603b56fb0a68642657602d6",
 			"permissioned_candidates_policy_id": "58b4ba68f641d58f7f1bba07182eca9386da1e88a34d47a14638c3fe",
+			"native_token": {
+				"asset": {
+					"policy_id": "ada83ddd029614381f00e28de0922ab0dec6983ea9dd29ae20eef9b4",
+					"asset_name": "5043546f6b656e44656d6f",
+				},
+				"illiquid_supply_address": "addr_test1wqn2pkvvmesmxtfa4tz7w8gh8vumr52lpkrhcs4dkg30uqq77h5z4"
+			},
 		},
 		"initial_permissioned_candidates": [
 			{
@@ -365,7 +367,7 @@ fn test_chain_config_content() -> serde_json::Value {
 			  "grandpa_pub_key": "0xd17c2d7823ebf260fd138f2d7e27d114c0145d968b5ff5006125f2414fadae69",
 			  "sidechain_pub_key": "0x0390084fdbf27d2b79d26a4f13f0ccd982cb755a661969143c37cbc49ef5b91f27"
 			}
-		]
+		],
 	})
 }